result added to UI

+++ shared/src/commonMain/kotlin/org/m4xvel/weatherapp/data/remote/WeatherResponse.kt
@@ -17,6 +17,6 @@ data class Wind(
 @Serializable
 data class Main (
     val temp: Double,
-    val humidity: Long,
-    val pressure: Long
+    val humidity: Int,
+    val pressure: Int
 )
