+++ shared/src/commonMain/kotlin/org/m4xvel/weatherapp/domain/model/Weather.kt
@@ -4,6 +4,6 @@ data class Weather(
     val name: String,
     val temp: Double,
     val speed: Double,
-    val humidity: Long,
-    val pressure: Long
+    val humidity: Int,
+    val pressure: Int
 )
